requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
DAWG-Python==0.7.2
DAWG2-Python==0.9.0
docopt==0.6.2
easyocr==1.7.2
et_xmlfile==2.0.0
filelock==3.18.0
fsspec==2025.5.1
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
ninja==1.11.1.4
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pdf2image==1.17.0
pillow==11.2.1
pyclipper==1.3.0.post6
pymorphy2-dicts-ru==2.4.417127.4579844
pymorphy3==2.0.3
pymorphy3-dicts-ru==2.4.417150.4580142
PyMuPDF==1.26.0
pyspellchecker==0.8.3
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rutermextract==0.3
scikit-image==0.25.2
scipy==1.15.3
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
sympy==1.14.0
tifffile==2025.5.26
torch==2.7.0
torchvision==0.22.0
triton==3.3.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
